Se modularizo el codigo
